requirements.txt
asn1crypto==0.24.0
cffi==1.11.2
idna==2.6
pycparser==2.18
pycryptodome==3.4.7
six==1.11.0
